First Push from Paul- Scans late2022 early2023

+++ SignalBiasScan/Analysis/VoltageSteps/Code/VoltageStepsMaker.h
@@ -23,7 +23,7 @@ class VoltageStepsMaker {
    // Returns voltage from evt number or timestamp
    int getVoltage_evtnumber(int run, int evt, int timestamp=0);
    
-  int getVoltage_timestamp(double timestamp); //Changed
+  int getVoltage_timestamp(int timestamp); //Changed
    //int getVoltage_timestamp(int timestamp); //FIXME
 
    // Returns voltage step index
